Refactor into scene manager and class
