+++ frontend/src/app/app.module.ts
@@ -5,7 +5,6 @@ import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { SceneRunnerComponent } from './scene-runner/scene-runner.component';
 import { BaseSceneComponent } from './base-scene/base-scene.component';
-import { FirstSceneComponent } from './scenes/first-scene/first-scene.component';
 import { SceneControllerComponent } from './scene-controller/scene-controller.component';
 import { SceneManagerComponent } from './scene-manager/scene-manager.component';
 
@@ -14,7 +13,6 @@ import { SceneManagerComponent } from './scene-manager/scene-manager.component';
     AppComponent,
     SceneRunnerComponent,
     BaseSceneComponent,
-    FirstSceneComponent,
     SceneControllerComponent,
     SceneManagerComponent
   ],
