+++ frontend/src/app/scene-manager/scene-manager.component.html
@@ -1,1 +1,15 @@
-<app-first-scene></app-first-scene>
+<div class="screen">
+  <div class="story">
+    <canvas #canvas class="layer"></canvas>
+    <svg class="layer" width="1000" height="810">
+      <g>
+        <rect class="hover" x="500" y="400" width="30" height="10" />
+      </g>
+    </svg>
+  </div>
+</div>
+
+
+<div style="display: none;">
+  <img id="handImage" src="assets/71434d054504aa872973bc782550f1a2.png">
+</div>
