+++ frontend/src/app/scene-manager/scene-manager.component.scss
@@ -1,0 +1,26 @@
+div.screen {
+  height: 900vh;
+  width: 100%;
+}
+
+div.story {
+  position: fixed;
+}
+
+canvas {
+  height: 100vh;
+  width: 100vw;
+}
+
+.layer {
+  position: absolute;
+}
+
+.hover {
+  fill: none;
+  pointer-events: all;
+}
+
+.hover:hover {
+  fill: red;
+}
