+++ frontend/src/app/scene-runner/scene-runner.component.ts
@@ -1,5 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 
+
+/*
+ * SceneRunner is responsible of drawing a scene
+ */
 @Component({
   selector: 'app-scene-runner',
   templateUrl: './scene-runner.component.html',
