+++ frontend/src/app/scene.ts
@@ -1,3 +1,14 @@
-export class Scene {
+export abstract class Scene {
 
+  numFrame: number;
+
+  constructor(
+    protected ctx: CanvasRenderingContext2D,
+    protected size: {height: number; width: number},
+    numFrame: number,
+  ) {
+    this.numFrame = numFrame;
+  }
+
+  abstract draw(frameNum: number): void;
 }
