--- frontend/src/app/scenes/first-scene/first-scene.component.html
@@ -1,14 +1,0 @@
-<div class="screen">
-  <div class="story">
-    <canvas #canvas class="layer"></canvas>
-    <svg class="layer" width="1000" height="810">
-      <g>
-        <rect class="hover" x="500" y="400" width="30" height="10" />
-      </g>
-    </svg>
-  </div>
-</div>
-
-<div style="display: none;">
-  <img #hand src="assets/71434d054504aa872973bc782550f1a2.png">
-</div>
