--- frontend/src/app/scenes/first-scene/first-scene.component.scss
@@ -1,26 +1,0 @@
-div.screen {
-  height: 900vh;
-  width: 100%;
-}
-
-div.story {
-  position: fixed;
-}
-
-canvas {
-  height: 100vh;
-  width: 100vw;
-}
-
-.layer {
-  position: absolute;
-}
-
-.hover {
-  fill: none;
-  pointer-events: all;
-}
-
-.hover:hover {
-  fill: red;
-}
